// VerFinalRevisionPizcun/src/app/interfaces/consultas.interface.ts
export interface Consultas { 
  id_con?:            string;
  id_ciu?:            string;
  nombre_ciu?:        string;
  id_sede?:           string;
  nombre_sede?:       string;
  id_ma?:             string;
  nombre_ma?:         string;
  id_equi?:           string;
  nombre_equi?:       string;
  id_pro?:            string;
  nombre_pro?:        string;
  id_ubi?:            string;
  nombre_ubi?:        string;
  modelo_con?:        string;
  serial_con?:        string;
  placa_con?:         string;
  mantenimiento_con?: string;
  status?: string;
  code?:   number;
  message?: string;
}